Fixes coordinate system and tests it to keep it that way
We decided to index into the matrix, be it energy or image, by
mathematical notation; thus, [0][1] corresponds to the x = 0, y = 1
spot in the matrix.  We have adopted the convention of using rank to
describe this as follows: RANK_ROW_COLUMN.  e.g. a matrix with x
dimension of 2 and y dimension of 3 is RANK_2_3.
 - src/createStartingPoints.js:createStartingPointValues swapped
   references between rows and columns
 - src/findSeams.js:findSeams protected against empty energy map
 - src/findSeams.js:createBackTracker corrected directions to
   backtrack in
 - src/findSeams.js:startCoord corrected coord system for last coord
 - src/findSeams.js:backTrackSeam flipped dimensions
 - src/findSeams.js:createEmptyDistanceCache flipped dimensions
 - src/findSeams.js:executeAlgorithm flipped dimensions
 - src/findSeams.js:append flipped dimensions